refactor(global id): refactor node module into the global id module
the global id module then implements the global id spec
