+++ src/global-object-identification/global-id.field.ts
