+++ src/global-object-identification/global-id.scalar.spec.ts
