+++ src/global-object-identification/global-id.scalar.ts
