+++ src/global-object-identification/global-id.type.ts
