+++ src/global-object-identification/index.ts
@@ -2,8 +2,6 @@ export { GlobalIdFieldResolver } from './global-id-field.resolver';
 export { GlobalIdField, GlobalIdFieldOptions } from './global-id.field';
 export { GlobalIdScalar } from './global-id.scalar';
 export { GlobalId } from './global-id.type';
-export {
-  ResolvedGlobalId,
-  typeResolvedGlobalId,
-  typeResolvedGlobalIds,
-} from './resolved-global-id.class';
+export { NodeInterface } from './node.interface';
+export { NodeFieldResolver, NodeResolver } from './node-field.resolver';
+export { ResolvedGlobalId } from './resolved-global-id.class';
