+++ src/global-object-identification/node.interface.spec.ts
