+++ src/global-object-identification/resolved-global-id.class.spec.ts
