+++ src/global-object-identification/resolved-global-id.class.ts
