+++ src/nestjs-relay.ts
@@ -1,3 +1,2 @@
-export * from './global-id';
+export * from './global-object-identification';
 export * from './mutation';
-export * from './node';
