--- src/node/index.ts
@@ -1,2 +1,0 @@
-export { NodeInterface } from './node.interface';
-export { NodeFieldResolver, NodeResolver } from './node-field.resolver';
